# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Discord Bot JDA API
  description: Complete API for interacting with a Java Discord bot using JDA
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /guilds/{guildId}:
    get:
      summary: Get guild information
      description: Returns detailed information about a specific guild
      operationId: getGuild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed guild information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
        '404':
          description: Guild not found
  /guilds/{guildId}/channels:
    get:
      summary: Get guild channels
      description: Returns all channels in a specific guild
      operationId: getGuildChannels
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of channels in the guild
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
        '404':
          description: Guild not found
  /guilds/{guildId}/members:
    get:
      summary: Get guild members
      description: Returns members of a specific guild
      operationId: getGuildMembers
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: after
          in: query
          required: false
          schema:
            type: string
            description: Get members after this user ID
      responses:
        '200':
          description: A list of members in the guild
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GuildMember'
        '404':
          description: Guild not found
...
  /guilds/{guildId}/roles:
    get:
      summary: Get guild roles
      description: Returns all roles in a specific guild
      operationId: getGuildRoles
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of roles in the guild
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '404':
          description: Guild not found
  /guilds/{guildId}/emojis:
    get:
      summary: Get guild emojis
      description: Returns all emojis in a specific guild
      operationId: getGuildEmojis
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of emojis in the guild
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emoji'
        '404':
          description: Guild not found
...
components:
  schemas:
...
    Guild:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Discord ID of the guild
        name:
          type: string
          description: Name of the guild
        icon:
          type: string
          format: uri
          description: URL of the guild icon
        iconHash:
          type: string
          description: Hash of the guild icon
        splash:
          type: string
          description: Hash of the guild splash image
        discoverySplash:
          type: string
          description: Hash of the guild discovery splash image
        owner:
          type: boolean
          description: Whether the bot is the owner of the guild
        ownerId:
          type: string
          description: ID of the guild owner
        permissions:
          type: string
          description: Permissions for the bot in the guild (as a bit set)
        region:
          type: string
          description: Voice region of the guild
        afkChannelId:
          type: string
          description: ID of the AFK channel
        afkTimeout:
          type: integer
          description: AFK timeout in seconds
        widgetEnabled:
          type: boolean
          description: Whether the widget is enabled
        widgetChannelId:
          type: string
          description: Channel ID for the guild widget
        verificationLevel:
          type: integer
          description: Verification level required for the guild
        defaultMessageNotifications:
          type: integer
          description: Default message notification level
        explicitContentFilter:
          type: integer
          description: Explicit content filter level
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          description: Roles in the guild
        emojis:
          type: array
          items:
            $ref: '#/components/schemas/Emoji'
          description: Custom emojis in the guild
        features:
          type: array
          items:
            type: string
          description: Enabled guild features
        mfaLevel:
          type: integer
          description: Required MFA level for the guild
        applicationId:
          type: string
          description: Application ID of the guild creator if bot-created
        systemChannelId:
          type: string
          description: ID of the system channel
        systemChannelFlags:
          type: integer
          description: System channel flags
        rulesChannelId:
          type: string
          description: ID of the rules channel
        joinedAt:
          type: string
          format: date-time
          description: When the bot joined the guild
        large:
          type: boolean
          description: Whether this is considered a large guild
        unavailable:
          type: boolean
          description: Whether the guild is unavailable
        memberCount:
          type: integer
          description: Total number of members in the guild
        voiceStates:
          type: array
          items:
            $ref: '#/components/schemas/VoiceState'
          description: Voice states of members in the guild
        members:
          type: array
          items:
            $ref: '#/components/schemas/GuildMember'
          description: Members in the guild
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
          description: Channels in the guild
        threads:
          type: array
          items:
            $ref: '#/components/schemas/ThreadChannel'
          description: Threads in the guild
        presences:
          type: array
          items:
            $ref: '#/components/schemas/Presence'
          description: Presences of the members in the guild
        maxPresences:
          type: integer
          description: Maximum number of presences for the guild
        maxMembers:
          type: integer
          description: Maximum number of members for the guild
        vanityUrlCode:
          type: string
          description: Vanity URL code for the guild
        description:
          type: string
          description: Description of the guild
        banner:
          type: string
          description: Banner hash
        premiumTier:
          type: integer
          description: Premium tier
        premiumSubscriptionCount:
          type: integer
          description: Number of boosts this guild currently has
        preferredLocale:
          type: string
          description: Preferred locale
        publicUpdatesChannelId:
          type: string
          description: ID of the public updates channel
        maxVideoChannelUsers:
          type: integer
          description: Maximum amount of users in a video channel
        approximateMemberCount:
          type: integer
          description: Approximate number of members in this guild
        approximatePresenceCount:
          type: integer
          description: Approximate number of presences in this guild
        welcomeScreen:
          type: object
          description: Welcome screen object
        nsfwLevel:
          type: integer
          description: NSFW level
        stickers:
          type: array
          items:
            $ref: '#/components/schemas/Sticker'
          description: Custom stickers in the guild
    GuildMember:
      type: object
      required:
        - user
        - joinedAt
      properties:
        user:
          $ref: '#/components/schemas/User'
          description: The user this guild member represents
        nick:
          type: string
          description: Guild nickname
        avatar:
          type: string
          description: Guild avatar hash
        roles:
          type: array
          items:
            type: string
          description: Role IDs
        joinedAt:
          type: string
          format: date-time
          description: When the user joined the guild
        premiumSince:
          type: string
          format: date-time
          description: When the user started boosting the guild
        deaf:
          type: boolean
          description: Whether the user is deafened in voice channels
        mute:
          type: boolean
          description: Whether the user is muted in voice channels
        pending:
          type: boolean
          description: Whether the user has not yet passed the guild's Membership Screening requirements
        permissions:
          type: string
          description: Guild permissions
        communicationDisabledUntil:
          type: string
          format: date-time
          description: When the user's timeout will expire
...
    Channel:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          description: Discord ID of the channel
        type:
          type: integer
          description: Type of the channel
          enum: [0, 1, 2, 3, 4, 5, 6, 10, 11, 12, 13, 14, 15]
        guildId:
          type: string
          description: ID of the guild this channel belongs to
        position:
          type: integer
          description: Position of the channel
        permissionOverwrites:
          type: array
          items:
            $ref: '#/components/schemas/PermissionOverwrite'
          description: Permission overwrites for the channel
        name:
          type: string
          description: Name of the channel
        topic:
          type: string
          description: Topic of the channel
        nsfw:
          type: boolean
          description: Whether the channel is NSFW
        lastMessageId:
          type: string
          description: ID of the last message sent in this channel
        bitrate:
          type: integer
          description: Bitrate of a voice channel
        userLimit:
          type: integer
          description: User limit of a voice channel
        rateLimitPerUser:
          type: integer
          description: Slow mode duration in seconds
        recipients:
          type: array
          items:
            $ref: '#/components/schemas/User'
          description: Recipients of a DM
        icon:
          type: string
          description: Icon hash
        ownerId:
          type: string
          description: ID of the DM creator
        applicationId:
          type: string
          description: ID of the application that created this channel
        parentId:
          type: string
          description: ID of the parent category for this channel
        lastPinTimestamp:
          type: string
          format: date-time
          description: When the last pinned message was pinned
        rtcRegion:
          type: string
          description: Voice region of the channel
        videoQualityMode:
          type: integer
          description: Video quality mode of the channel
        messageCount:
          type: integer
          description: Number of messages in a thread
        memberCount:
          type: integer
          description: Number of members in a thread
        threadMetadata:
          $ref: '#/components/schemas/ThreadMetadata'
          description: Thread-specific fields
        member:
          $ref: '#/components/schemas/ThreadMember'
          description: Thread member for the current user
        defaultAutoArchiveDuration:
          type: integer
          description: Default auto archive duration for threads
        permissions:
          type: string
          description: Computed permissions for the invoking user
        flags:
          type: integer
          description: Channel flags
        totalMessageSent:
          type: integer
          description: Number of messages ever sent in a thread
        availableTags:
          type: array
          items:
            $ref: '#/components/schemas/ForumTag'
          description: Available tags for a forum channel
        appliedTags:
          type: array
          items:
            type: string
          description: IDs of tags applied to a forum thread
        defaultReactionEmoji:
          $ref: '#/components/schemas/DefaultReaction'
          description: Default reaction for a forum channel
        defaultThreadRateLimitPerUser:
          type: integer
          description: Default slow mode for forum threads
        defaultSortOrder:
          type: integer
          description: Default sort order for forum posts
        defaultForumLayout:
          type: integer
          description: Default forum layout view
...
    Emoji:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Emoji ID
        name:
          type: string
          description: Emoji name
        roles:
          type: array
          items:
            type: string
          description: Roles allowed to use this emoji
        user:
          $ref: '#/components/schemas/User'
          description: User that created this emoji
        requireColons:
          type: boolean
          description: Whether this emoji must be wrapped in colons
        managed:
          type: boolean
          description: Whether this emoji is managed
        animated:
          type: boolean
          description: Whether this emoji is animated
        available:
          type: boolean
          description: Whether this emoji can be used

    # Sticker related schemas
...